// apps/vendor/src/services/models/TransactionViewPagedCollection.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { Link } from './Link';
import type { TransactionView } from './TransactionView';

export type TransactionViewPagedCollection = {
    offset?: number | null;
    limit?: number | null;
    size?: number;
    first?: Link;
    previous?: Link;
    next?: Link;
    last?: Link;
    self?: Link;
    value?: Array<TransactionView> | null;
};